Remove SharedVariableMap from server session
As part of the move toward a stateless server session, the m_sharedVariableMap has been removed from the server session.

The SharedVariableMap has been renamed to VariableMap and is now used exclusively on the client side.

Instead of synchronizing variables from server to client, the client session initializes its VariableMap using the ISessionService.

To support updates after initialization, the server can send a client notification. A corresponding handler on the client side should process the notification and update the relevant variables on the client session.

423389