Table: preserve scrollTop on replaceRows
When rows are replaced or during a table reload, the scroll position
 should be preserved.
In Java, this works by accident. The filler of the table data consumes
the space temporarily together with the JS scrollbars. However, it
does not work perfectly: the table sometimes scrolls about 8px on
reload and if the table is scrolled nearly at the end, a reload will
scroll it to the end. And most importantly, it does not work with
native scrollbars (e.g. on Safari).

In Scout JS, it does not work if a SmartColumn is used, because the
rendering of the viewport is blocked until the texts are resolved.
So the existing rows are removed, fillers reset to 0 but no new rows
rendered -> scrollTop will be set to 0.
The new solution just restores scrollTop after the rows are inserted
and when all the texts are resolved. This also works with native
scrollbars.

423551