aria: mark invalid fields and consider status for description and error
Fields with an error now get the attribute aria-invalid and
aria-errormessage which is linked to the content of the displayed error
status.

If a status is set, but it doesn't have severity ERROR, the message is
written into aria-description replacing the tooltip text if there is
one. This also happens visually: a status message replaces the tooltip.

394433