TreeAdapter: mark remote nodes
Tree node events sent to the UI server for nodes that do not exist on
the UI server create a lot of noise in the log. This may be acceptable
if Scout JS can not know if the UI server knows a tree node or not but
it is not ok if Scout JS can determine if a tree node does not exist on
the UI server.
Therefore, all tree nodes created by the TreeAdapter are now marked as
remote. This flag is used to filter events that are sent back to the UI
server. In addition, the filter method introduces hybrid nodes which are
nodes created as a wrapper in Java while all their logic is implemented
in Scout JS. The only known variant of these hybrid nodes are JsPages
and these are marked accordingly by the OutlineAdapter.

427023