Add child pages to IJsPage
An IJsPage is a page created in Java as a wrapper while all its logic is
implemented in Scout JS. As this page lives in a Scout Classic Outline
all its child pages need to be known to the Outline in Java as well.
Therefore, the implementation in Scout JS can use the new JsPageHelper
and call the server to create its child pages for a list of IIds or
IPageParamDos. The creation of the child pages is then delegated to the
IJsPage wrapper in Java. Afterward, the JsPageHelper holds references to
the created child pages that can be used by the Scout JS page again.
The created child pages will be identified by an IPageParamDo and they
are inserted into the Outline directly on the server. The JsPageHelper
ensures that necessary information is synchronized, like e.g. summary
information from a parent table page.
The tree will now send nodeChanged events back to the server in order to
update text, htmlEnabled, cssClass and iconId if it was changed e.g. by
a parent PageWithTable.
If there are childNodes present on the server for an IJsPage these
childNodes and the properties expanded and expandedLazy are added to the
jsPageModel. This ensures the outline state being correct even after a
browser reload.

427023