aria: make table areas tabbable
Table controls and header are now tabbable. It works the same way as
for menu bars: pressing tab focuses the first item, the next item can
be focused using arrow right key.

The logic from the menubar has been extracted to a coordinator to make
it reusable. The coordinator can handle actions but also plain jquery
items which is necessary for the table header.

Also:
- BreadcrumbMenuBar is now tabbable and uses the same principle
- The filter icons in the table header menu are now tabbable. They were
  converted to actions to get the toggle keystroke for free. Their
  styling has also been improved: they now have a proper hover effect
  and the same size. Flex-box is used to get a proper DOM order which
  is not possible using float: right.

394500